Initial code commit.
This is a full rewrite of the boot loader to bring it up to date and current with latest client. I took the time to do a full rewrite and overhaul of this loader because it was much needed to be in a more maintainable state for the project. There is a lot of debug information on our forums which will be released soon that I took the time to create and explain how this loader works and is created.

Please use our forums for more information regarding this project.
You may use the issues/wiki on Github to report problems with this loader.